requirements.txt
# Web Framework
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary
alembic

# Task Queue
celery
redis

# Environment & Configuration
python-dotenv

# File Upload
python-multipart

# Templates (for frontend)
jinja2

